Add `disableProcessIdCheck` setting for Docker compatibility.
ANCM verifies that the process *it* started (or one of its children) is actually listening on `ASPNETCORE_PORT`,
and considers the process launch failed if it finds a mismatch.

This is a problem for Docker on Windows; the
`docker.exe` user-mode program *triggers* a containerized app to start, but the `com.docker.service` system
service actually hosts the application, and opens the ports it exposes.

The opt-in `disableProcessIdCheck` setting relaxes the ANCM port check to only verify that *some* process is
listening on the expected port (not necessarily *the* process it started or its children). With this change, ANCM
can successfully launch and proxy requests to Dockerized applications running on Windows.

A separate repository (https://github.com/nil4/TestANCM) includes sample applications that run sucessfully
with the modified ANCM behind IIS Express, using Windows authentication:
- NodeJS and Python running natively on Windows, and implementing the ANCM pairing token protocol
- a simple Go application (HTTP echo server) running in a LCOW container on Windows
- a .NET Core application deployed to an Alpine container